bugfix: switching GPUs
closes #110
